requirements.txt
flask>=3.0
flask-cors>=4.0
PyPDF2>=3.0
openai>=1.0
python-dotenv>=1.0
numpy>=1.24
gunicorn>=21.2
